pyproject.toml
[project]
name = "dictation-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.23.0",
    "pyaudio>=0.2.14",
    "pynput>=1.8.1",
    "pyobjc-core>=12.1",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
    "quickmachotkey>=2025.7.28",
]
